+++ lambdas/indexers/dmp/app.rb
@@ -229,6 +229,8 @@ module Functions
 
       # Fetch the index record from the DB
       def _dynamo_index_get(client:, table:, key:, logger: nil)
+        return JSON.parse(key.to_json) if key[:PK].nil? || key[:SK].nil?
+
         resp = client.get_item({
           table_name: table,
           key:,
@@ -243,6 +245,8 @@ module Functions
 
       # Add/update an index record
       def _dynamo_index_put(client:, table:, item:, logger: nil)
+        return false if item[:PK].nil? || item[:SK].nil?
+
         resp = client.put_item({
           table_name: table,
           item:,
@@ -259,35 +263,92 @@ module Functions
 
         project = hash.fetch('project', {}).fetch('L', [{}]).first['M']
         project = {} if project.nil?
-        # Set the project start date equal to the date specified or the DMP creation date
-        proj_start = project.fetch('start', hash.fetch('created', {}))['S']
-        # Set the project end date equal to the specified end OR 5 years after the start
-        proj_end = project.fetch('end', {})['S']
-        proj_end = (Date.parse(proj_start.to_s) + 1825).to_s if proj_end.nil? && !proj_start.nil?
+        # Calculate the project start and end dates
+        project_dates = _getProjectDates(project: project)
 
         funding = project.fetch('funding', {}).fetch('L', [{}]).first.fetch('M', {})
         doc = people.merge(_extract_funding(hash: funding, logger:))
         doc = doc.merge(_repos_to_os_doc_parts(datasets: hash.fetch('dataset', {}).fetch('L', [])))
+
         doc = doc.merge({
           dmp_id: Uc3DmpId::Helper.remove_pk_prefix(p_key: pk),
-          title: hash.fetch('title', {})['S']&.downcase,
-          visibility: visibility,
+          title: _remove_markup(text: hash.fetch('title', {})['S'])&.gsub(/'"/, '')&.to_s,
+          visibility: visibility&.to_s,
           featured: hash.fetch('dmproadmap_featured', {})['S']&.downcase&.strip == '1' ? 1 : 0,
-          description: hash.fetch('description', {})['S']&.downcase,
-          project_start: proj_start&.to_s&.split('T')&.first,
-          project_end: proj_end&.to_s&.split('T')&.first,
-          created: hash.fetch('created', {})['S']&.to_s&.split('T')&.first,
-          modified: hash.fetch('modified', {})['S']&.to_s&.split('T')&.first,
-          registered: hash.fetch('registered', {})['S']&.to_s&.split('T')&.first
+          description: _remove_markup(text: hash.fetch('description', {})['S'])&.gsub(/'"/, '')&.to_s,
+          project_start: project_dates[:project_start]&.to_s,
+          project_end: project_dates[:project_end]&.to_s,
+          created: hash.fetch('created', {})['S']&.to_s&.split('T')&.first&.to_s,
+          modified: hash.fetch('modified', {})['S']&.to_s&.split('T')&.first&.to_s,
+          registered: hash.fetch('registered', {})['S']&.to_s&.split('T')&.first&.to_s
         })
         logger.debug(message: 'New OpenSearch Document', details: { document: doc }) unless visibility == 'public'
         return doc unless visibility == 'public'
 
         # Attach the narrative PDF if the plan is public
         works = hash.fetch('dmproadmap_related_identifiers', hash.fetch('dmproadmap_related_identifier', {})).fetch('L', [])
-        doc[:narrative_url] = _extract_narrative(works:, logger:)
+        doc[:narrative_url] = _extract_narrative(works:, logger:)&.to_s
         logger.debug(message: 'New OpenSearch Document', details: { document: doc })
         doc
+      end
+
+      # Fetch the project start and end dates
+      def _getProjectDates(project:)
+        three_years = 365 * 3
+
+        begin
+          start_date = Date.parse(project.fetch('start', project.fetch('created', {}))['S'])
+        rescue
+          start_date = nil
+        end
+
+        begin
+          end_date = Date.parse(project.fetch('end', {})['S'])
+        rescue
+          end_date = nil
+        end
+
+        if start_date.nil?
+          # Start date should be 3 years prior to the end date if it was not provided
+          start_date = end_date - three_years if !end_date.nil?
+        else
+          # End date should be 3 years from the start date if it was not provided or is less than 3 years
+          three_years_out = start_date + three_years unless start_date.nil?
+          end_date = three_years_out if end_date.nil? || end_date < three_years_out
+        end
+
+        { project_start: start_date&.strftime('%Y-%m-%d'), project_end: end_date&.strftime('%Y-%m-%d') }
+      end
+
+      # Check if the value is a valid date
+      def _isDate(val:)
+        regex = /[\d]{4}\-[\d]{2}\-[\d]{2}/
+        return false if !val.is_a?(String) || (val =~ regex).nil?
+
+        parts = val.split('-')
+        return false if parts[0].length != 4 || parts[0].to_i < 2000
+        return false if parts[1].length != 2 || parts[1].to_i < 1 || parts[1].to_i > 12
+        # I realize that not every month has 31 days, this should be enough though to catch most issues
+        return false if parts[2].length != 2 || parts[2].to_i < 1 || parts[2].to_i > 31
+
+        true
+      end
+
+      # Remove any HTML markup fropm the text
+      def _remove_markup(text:)
+        return nil if text.nil? || text.strip.blank?
+
+        regex = /<\/?[a-zA-Z0-9\s="'\-_]\/?>/
+        text.split(regex).reject { |e| e.gsub("\\r", '').gsub("\\n", ' ').strip == '' }.join(' ')
+      end
+
+      # Remove the parenthesis part of the name and then convert to lowercase
+      def _prep_org_name(text:)
+        return nil if text.nil? || text.strip.blank?
+
+        regex = /\s+\(.+\)$/
+        out = text.gsub(regex, '').gsub(/'"/, '').strip
+        out.blank? ? nil : out.downcase
       end
 
       # Extract the important funding info
@@ -295,16 +356,34 @@ module Functions
         return {} unless hash.is_a?(Hash)
 
         id = hash.fetch('funder_id', {}).fetch('M', {}).fetch('identifier', {})['S']
+        id = id&.gsub('https://doi.org/10.13039/', '')&.gsub('https://ror.org/', '')&.to_s
         {
-          funder_ids: [id].compact.uniq,
-          funders: [hash.fetch('name', {})['S']&.downcase&.strip].compact.uniq,
+          funder_ids: [id],
+          funders: [_prep_org_name(text: hash.fetch('name', {})['S'])],
           funder_opportunity_ids: [
             hash.fetch('dmproadmap_funding_opportunity_id', {})['S']&.downcase&.strip,
             hash.fetch('dmproadmap_project_number', {})['S']&.downcase&.strip
           ].compact.uniq,
-          grant_ids: [hash.fetch('grant_id', {}).fetch('M', {}).fetch('identifier', {})['S']].compact.uniq,
-          funding_status: hash.fetch('funding_status', {}).fetch('S', 'planned')
+          grant_ids: [_process_grant_id(id: hash.fetch('grant_id', {}).fetch('M', {}).fetch('identifier', {})['S'])],
+          funding_status: hash.fetch('funding_status', {}).fetch('S', 'planned')&.to_s
         }
+      end
+
+      # Remove any URL prefixes from the grant id
+      def _process_grant_id(id:)
+        # example NSF Award URL: https://www.nsf.gov/awardsearch/showAward?AWD_ID=0000000&HistoricalAwards=false
+        nsf_awd_regex = /AWD_ID=([0-9]+)/
+        out = nsf_awd_regex.match(id).to_s
+        return out.gsub('AWD_ID=', '') unless out.blank?
+
+        # example NSF URL with NSF ID: https://www.nsf.gov/publications/pub_summ.jsp?org=NSF&ods_key=nsf00000
+        nsf_id_regex = /nsf[0-9\-]+/
+        out = nsf_id_regex.match(id).to_s
+        return out.gsub('nsf', '') unless out.blank?
+
+        # otherwise just take the first 5 digit portion or return it as-is
+        out = /\d{5,}/.match(id).to_s
+        out.blank? ? id : out
       end
 
       # Extarct the latest link to the narrative PDF
@@ -364,11 +443,17 @@ module Functions
           hosts.each do |host|
             next if host.nil?
 
-            parts[:repos] << host.fetch('title', {})['S']
-            parts[:repo_ids] << host.fetch('url', {})['S']
-            parts[:repo_ids] << host.fetch('dmproadmap_host_id', {}).fetch('M', {}).fetch('identifier', {})['S']
+            parts[:repos] << host.fetch('title', {})['S']&.to_s
+            parts[:repo_ids] << host.fetch('url', {})['S']&.to_s
+
+            re3url = 'https://www.re3data.org/api/v1/repository/'
+            host_id = host.fetch('dmproadmap_host_id', {}).fetch('M', {}).fetch('identifier', {})['S']&.to_s
+            parts[:repo_ids] << host_id
+            # Include a cn entry for the re3data id without the full URL
+            parts[:repo_ids] << host_id.gsub(re3url, '') if host_id.start_with?(re3url)
           end
         end
+
         parts[:repo_ids] = parts[:repo_ids].compact.uniq
         parts[:repos] = parts[:repos].compact.uniq
         parts
@@ -405,9 +490,11 @@ module Functions
         id_type = hash['M']['contact_id'].nil? ? 'contributor_id' : 'contact_id'
         affiliation = _process_affiliation(hash: hash['M'].fetch('dmproadmap_affiliation', {}))
 
+        name = hash['M'].fetch('name', {})['S']&.downcase&.gsub(/\s+/, ' ')
+        name = name.split(', ').reverse.join(' ') unless name.nil?
         {
-          name: hash['M'].fetch('name', {})['S']&.downcase,
-          email: hash['M'].fetch('mbox', {})['S']&.downcase,
+          name: hash['M'].fetch('name', {})['S']&.downcase&.gsub(/\s+/, ' ')&.to_s,
+          email: hash['M'].fetch('mbox', {})['S']&.downcase&.to_s,
           id: _process_id(hash: hash['M'].fetch(id_type, {})),
           affiliation: affiliation[:name],
           affiliation_id: affiliation[:id]
@@ -429,8 +516,8 @@ module Functions
         return {} unless hash.is_a?(Hash) && !hash['M'].nil?
 
         {
-          name: hash['M'].fetch('name', {})['S']&.downcase,
-          id: _process_id(hash: hash['M'].fetch('affiliation_id', {}))
+          name: _prep_org_name(text: hash['M'].fetch('name', {})['S'])&.to_s,
+          id: _process_id(hash: hash['M'].fetch('affiliation_id', {}))&.to_s
         }
       end
 
@@ -441,7 +528,10 @@ module Functions
       #      "type": { "S": "ror" }
       #    }
       def _process_id(hash:)
-        hash.is_a?(Hash) ? hash.fetch('M', {}).fetch('identifier', {})['S']&.downcase : nil
+        return nil unless hash.is_a?(Hash)
+
+        id = hash.fetch('M', {}).fetch('identifier', {})['S']&.to_s
+        (id.nil? || id.blank?) ? nil : id.gsub('https://ror.org/', '').gsub('https://orcid.org/', '')
       end
     end
   end