+++ lambdas/layers/dmptool/__tests__/setup.ts
@@ -1,0 +1,1 @@
+// Setup env variables and any other things that we would need to mock for our tests